# feat(compound-interest): Enhance Readability and Efficency of compounding, utilizing modern, standard JS Syntax.