Old Features:

- Old compound-interest Function was written using a for function and if/else statements for the mulitplicators.
- The pre-defined array has had no length definition, therefore JS needs to dynamically allocate / reallocate memory while pushing elements into the array.
- The general code approach was semi-easy to read and understand, however for non-js/-ts developers easier to understand than for js-ts developers
- For functions are suboptimal, optimized by the javascript engine, the only best optimized for function would be a reverse function from end-to-start which makes no sense for a compounding function
- The main function has had too many lines for beeing a subroutine (definition of a function)

New Code:
- The newly rewritten code, does not use a for function instead uses the standard, well known, javascript array iteration function "reduce" to reduce the array into a combined object
- The acc. addition of temp variables was removed and we now just extract the last value of the array for the "last values"
- There are 2 subfunctions introduced, allowing the Easier navigation through code, which can later be extracted into a util area if wished
- The definition for the "duration, payment, interest" base investment data, is now done throug hthe type compounditnerestinput type input by object lookup, which removes overhead
- The new code is enhancing execution efficency by round about 35%


Here is a table of benchmarks:
| Input | Total Interest         | Avg Time Old | Avg Time New | Δ Avg (ms) | Min Time Old | Min Time New | Δ Min (ms) | Max Time Old | Max Time New | Δ Max (ms) |
|-------|------------------------|--------------|--------------|------------|--------------|--------------|------------|---------------|---------------|-------------|
| #0    | 63.181 €               | 0.180        | 0.116        | -0.064     | 0.155        | 0.103        | -0.052     | 12.216        | 11.619        | -0.597      |
| #1    | 2.297.928 €            | 0.176        | 0.120        | -0.056     | 0.157        | 0.108        | -0.049     | 1.071         | 1.213         | +0.142      |
| #2    | 193.573.644 €          | 0.184        | 0.135        | -0.049     | 0.161        | 0.118        | -0.043     | 1.857         | 1.092         | -0.765      |
| #3    | 189.619 €              | 0.172        | 0.118        | -0.054     | 0.154        | 0.105        | -0.049     | 1.130         | 1.101         | -0.029      |
| #4    | 34.413.062 €           | 0.175        | 0.126        | -0.049     | 0.157        | 0.110        | -0.047     | 1.777         | 1.564         | -0.213      |
| #5    | 268.088.808.015 €      | 0.182        | 0.135        | -0.047     | 0.162        | 0.118        | -0.044     | 2.138         | 1.670         | -0.468      |
| #6    | 1,5×10²⁵ €             | 0.192        | 0.134        | -0.058     | 0.164        | 0.117        | -0.047     | 12.530        | 11.297        | +1.233      |

Test for the benchmark was the following:
```ts
const testInputs: CompoundInterestInput[] = [
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 10,
        yearlyInterest: 0.07,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 40,
        yearlyInterest: 0.07,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 100,
        yearlyInterest: 0.07,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 10,
        yearlyInterest: 0.15,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 40,
        yearlyInterest: 0.15,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000,
        durationYears: 100,
        yearlyInterest: 0.15,
        type: 'monthly',
    },
    {
        startCapital: 10000,
        monthlyPayment: 1000000,
        durationYears: 100,
        yearlyInterest: 0.415,
        type: 'monthly',
    },
]

async function test() {
    const iterations = 25000

    for (let i = 0; i < testInputs.length; i++) {
        const input = testInputs[i]!
        const times: number[] = []

        let result = 'interest'
        // Run the test 100 times
        for (let j = 0; j < iterations; j++) {
            const start = performance.now()
            result = calcCompoundInterest(input).totalInterest
            const end = performance.now()
            times.push(end - start)
        }

        // Calculate average time
        const averageTime =
            times.reduce((sum, time) => sum + time, 0) / times.length
        const minTime = Math.min(...times)
        const maxTime = Math.max(...times)

        console.warn(`Input #${i}:`)
        console.warn(`  Average time: ${averageTime.toFixed(3)} milliseconds`)
        console.warn(`  Min time: ${minTime.toFixed(3)} milliseconds`)
        console.warn(`  Max time: ${maxTime.toFixed(3)} milliseconds`)
        console.warn(`  Total Interest: ${result}`)
        console.warn('')
    }
}

test()
```

I have tested it locally with bun and node-ts and tsx runtimes for typescript. On all runtimes similar results were granted.
In case, for beeing executed within the browser i have re-written it javascript using a strip-types approach and tested it in a browser console, where similar benchmarking results of efficency gain were seen.

![Uploading image.png…]()

The client side execution speed depends more on how "fast" the browser and "up to date" it is. For "low end computers" using javascript features is recommended (not using a for loop) [Best practice from multiple companies, like meta/react, google/angular and the classic: airbnb]

*I left the comments so that you can understand why and what the change contains, if wished i can remove the comment later on*

If you wish further optimizations such as this across the repository, i would be happy to do it. Just let me know.
I didn't want to do "everything" if it's not wanted.
Sincerely,
Chrissy8283


*Tested on a PC with intel Core i7 13700kf and DDR5 Memory with Windows 11 and both firefox as well as chrome based browsers*
*Second test was on a laptop with an intel Pentium*